requirements.txt
streamlit
yfinance
alpaca-py
python-dotenv
pytest
pytest-mock
